requirements.txt
streamlit[pdf]
toml
st-theme
-e ./pesuacademy-dev
lxml
python-dotenv
requests
cryptography
extra-streamlit-components